ats_bullet_selector/pyproject.toml
[project]
name = "ats-bullet-selector"
version = "0.1.0"
description = "ATS-aware bullet selection microservice for resume tailoring"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "httpx>=0.27.0",
    "groq>=0.11.0",
    "sentence-transformers>=3.0.0",
    "spacy>=3.7.0",
    "PuLP>=2.8.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[project.scripts]
ats-select = "ats_bullet_selector.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ats_bullet_selector"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
